Add: CODEGEN_AI_PROVIDER, CODEGEN_KEEP_VENV and CODEGEN_AI_MODEL envvars to configure the code generator [GS-149]. Add: ai_gpt_fn_index.py and ai_gpt_fn_tables.py example files to generate the LC Tools / GPT Functions [GS-149]. Add: Rhymes.ai Aria model [GS-152].

+++ scripts/code_generation/run_schema_generator.sh
@@ -14,27 +14,27 @@ source venv/bin/activate
 if [ -f requirements.txt ]; then
     pip install -r requirements.txt
 else
-    pip install ollama requests openai
+    pip install requests ollama openai groq
     pip freeze > requirements.txt
 fi
+
+if [ "${CODEGEN_AI_PROVIDER}" = "" ]; then
+    CODEGEN_AI_PROVIDER="${LANGCHAIN_DEFAULT_MODEL}"
+fi
+if [ "${CODEGEN_AI_PROVIDER}" = "" ]; then
+    echo "ERROR: CODEGEN_AI_PROVIDER not set"
+    exit 1
+fi
+
+if [ "${USER_INPUT}" = "" ]; then
+    USER_INPUT="./user_input_example.md"
+fi
+
 PARAMETERS=""
+
+# Ollama specific parameters
 if [ "${OLLAMA_BASE_URL}" != "" ]; then
     PARAMETERS="${PARAMETERS} --ollama_base_url ${OLLAMA_BASE_URL}"
-fi
-if [ "${LANGCHAIN_DEFAULT_MODEL}" = "ollama" ]; then
-    if [ "${OLLAMA_MODEL}" != "" ]; then
-        PARAMETERS="${PARAMETERS} --model ${OLLAMA_MODEL}"
-    fi
-fi
-if [ "${LANGCHAIN_DEFAULT_MODEL}" = "nvidia" ]; then
-    if [ "${NVIDIA_MODEL}" != "" ]; then
-        PARAMETERS="${PARAMETERS} --model ${NVIDIA_MODEL}"
-    fi
-fi
-if [ "${LANGCHAIN_DEFAULT_MODEL}" = "chat_openai" ]; then
-    if [ "${OPENAI_MODEL}" != "" ]; then
-        PARAMETERS="${PARAMETERS} --model ${OPENAI_MODEL}"
-    fi
 fi
 if [ "${OLLAMA_TEMPERATURE}" != "" ]; then
     PARAMETERS="${PARAMETERS} --temperature ${OLLAMA_TEMPERATURE}"
@@ -42,21 +42,56 @@ fi
 if [ "${OLLAMA_STREAM}" != "" ]; then
     PARAMETERS="${PARAMETERS} --stream ${OLLAMA_STREAM}"
 fi
+
+# Model specific parameters
+
+if [ "${CODEGEN_AI_MODEL}" != "" ]; then
+    PARAMETERS="${PARAMETERS} --model ${CODEGEN_AI_MODEL}"
+else
+    if [ "${CODEGEN_AI_PROVIDER}" = "ollama" ]; then
+        if [ "${OLLAMA_MODEL}" != "" ]; then
+            PARAMETERS="${PARAMETERS} --model ${OLLAMA_MODEL}"
+        fi
+    fi
+    if [ "${CODEGEN_AI_PROVIDER}" = "nvidia" ]; then
+        if [ "${NVIDIA_MODEL}" != "" ]; then
+            PARAMETERS="${PARAMETERS} --model ${NVIDIA_MODEL}"
+        fi
+    fi
+    if [ "${CODEGEN_AI_PROVIDER}" = "chat_openai" ]; then
+        if [ "${OPENAI_MODEL}" != "" ]; then
+            PARAMETERS="${PARAMETERS} --model ${OPENAI_MODEL}"
+        fi
+    fi
+    if [ "${CODEGEN_AI_PROVIDER}" = "openai" ]; then
+        if [ "${OPENAI_MODEL}" != "" ]; then
+            PARAMETERS="${PARAMETERS} --model ${OPENAI_MODEL}"
+        fi
+    fi
+    if [ "${CODEGEN_AI_PROVIDER}" = "groq" ]; then
+        if [ "${GROQ_MODEL}" != "" ]; then
+            PARAMETERS="${PARAMETERS} --model ${GROQ_MODEL}"
+        fi
+    fi
+    if [ "${CODEGEN_AI_PROVIDER}" = "aria" ]; then
+        if [ "${RHYMES_ARIA_MODEL}" != "" ]; then
+            PARAMETERS="${PARAMETERS} --model ${RHYMES_ARIA_MODEL}"
+        fi
+    fi
+fi
+
+# Other parameters
 if [ "${AGENTS_COUNT}" != "" ]; then
     PARAMETERS="${PARAMETERS} --agents_count ${AGENTS_COUNT}"
 fi
-if [ "${USER_INPUT}" = "" ]; then
-    USER_INPUT="./user_input_example.md"
-fi
-PARAMETERS="${PARAMETERS} --provider ${LANGCHAIN_DEFAULT_MODEL}"
-PARAMETERS="${PARAMETERS} --user_input ${USER_INPUT}"
-echo "LANGCHAIN_DEFAULT_MODEL: ${LANGCHAIN_DEFAULT_MODEL}"
-echo "NVIDIA_MODEL: ${NVIDIA_MODEL}"
+
 echo ""
-echo "Executing: python schema_generator.py ${PARAMETERS}"
+echo "Executing: python schema_generator.py --user_input \"${USER_INPUT}\" --provider ${CODEGEN_AI_PROVIDER} ${PARAMETERS}"
 pwd
 echo ""
-python schema_generator.py ${PARAMETERS}
+python schema_generator.py --user_input "${USER_INPUT}" --provider ${CODEGEN_AI_PROVIDER} ${PARAMETERS}
 
 deactivate
-rm -rf venv
+if [ "${CODEGEN_KEEP_VENV}" != "1" ]; then
+    rm -rf venv
+fi
